refactor : ButtonProps 타입 정의를 types 폴더로 분리
